swap lambda to new bot
